+++ packages/quantic/force-app/main/default/lwc/quanticPager/quanticPager.js
@@ -4,25 +4,46 @@ import {registerComponentForInit, initializeWithHeadless} from 'c/quanticHeadles
 import next from '@salesforce/label/c.quantic_Next';
 import previous from '@salesforce/label/c.quantic_Previous';
 
+/** @typedef {import("coveo").Pager} Pager */
+/** @typedef {import("coveo").SearchEngine} SearchEngine */
+
+/**
+ * The `QuanticPager` provides buttons that allow the end user to navigate through the different result pages.
+ * @example
+ * <c-quantic-pager engine-id={engineId} number-of-pages="4"></c-quantic-pager>
+ */
 export default class QuanticPager extends LightningElement {
+  /**
+   * The ID of the engine instance the component registers to.
+   * @api
+   * @type {string}
+   */
+  @api engineId;
+  /**
+   * Specifies how many page buttons to display in the pager.
+   * @api
+   * @type {number}
+   * @defaultValue 5
+   */
+  @api numberOfPages = 5;
+
   /** @type {number[]} */
   @track currentPages = [];
   /** @type {boolean}*/
   @track hasResults
 
-  /** @type {import("coveo").Pager} */
+  /** @type {Pager} */
   pager;
-  /** @type {()=> void} */
+  /** @type {Function} */
   unsubscribe;
-  /** @type {() => void} */
+  /** @type {Function} */
   unsubscribeSearchStatus;
   /** @type {boolean} */
   hasPrevious;
   /** @type {boolean} */
   hasNext;
+  /** @type {number} */
   currentPage = 1;
-  /** @type {string} */
-  @api engineId;
 
   labels = {
     next,
@@ -38,10 +59,14 @@ export default class QuanticPager extends LightningElement {
   }
 
   /**
-   * @param {import("coveo").SearchEngine} engine
+   * @param {SearchEngine} engine
    */
   initialize = (engine) => {
-    this.pager = CoveoHeadless.buildPager(engine);
+    this.pager = CoveoHeadless.buildPager(engine, {
+      options: {
+        numberOfPages: Number(this.numberOfPages),
+      }
+    });
     this.searchStatus = CoveoHeadless.buildSearchStatus(engine);
     this.unsubscribe = this.pager.subscribe(() => this.updateState());
     this.unsubscribeSearchStatus = this.searchStatus.subscribe(() => this.updateState());
@@ -81,5 +106,12 @@ export default class QuanticPager extends LightningElement {
 
   get previousDisabled() {
     return !this.hasPrevious;
+  }
+
+  get currentPagesObjects() {
+    return this.currentPages.map((page) => ({
+      number: page,
+      selected: page === this.currentPage
+    }));
   }
 }