+++ packages/quantic/force-app/main/default/lwc/quanticSvg/quanticSvg.js
@@ -6,8 +6,22 @@ import noResults from './noResults.html';
 // @ts-ignore
 import _default from './default.html';
 
+/**
+ * The `QuanticSvg` is used internally to display static svg images.
+ * @example
+ * <c-quantic-svg name="noResults"></c-quantic-svg>
+ */
 export default class QuanticSvg extends LightningElement {
-  /** @type {string} */
+  /**
+   * The name of the image to display.
+   * The options are:
+   *   - `'pageNotFound'`
+   *   - `'noResults'`
+   *   - `undefined (default)`
+   * @api
+   * @type {string}
+   * @defaultValue `undefined`
+   */
   @api name;
 
   render() {